// src/models/models.ts
export interface ICreditCardItem {
    brand: CCBrands;
    number: string;
    exp_month: number;
    exp_year: number;
    fname: string;
    lname: string;
    cvv: string;
    first6: string;
    country: string;
    last4: string;
    address_line1: string;
    address_line2: string;
    address_city: string;
    address_state: string;
    address_zip: string;
    address_country: string;
    amount: number
    processortype: CCProcessors;
    clientip: string;
    external_reference_id: string;
    external_customer_reference: string;
    cardid: string;
    authid: string;
    chargeid: string;
}

export interface IAuthorization {
    authid: string;
    amount: number;
    clientip: string;
    external_reference_id: string;
    processortype: CCProcessors;
}

export enum CCBrands {
    MC = 'MC',
    VISA = "VISA",
    AMEX = "AMEX",
    DISCOVER = "DISCOVER"
}

export class ChargeResult {

    // success = true is successfully charged
    success = false;

    // the raw response from the processor
    result: string = "";

    //message from processor
    message: string = "";

    // card number tokenized
    cardid: string = "";

    // authorization id from authorization
    authid: string = "";

    // chargeid from charge generated or transactionid generated by the processor
    chargeid: string = "";
}

export enum CCProcessors {
    clover = "clover",
    celero = "celero",
}

export interface Route {
    Path: string;
    FnCall: any;
    Authenticated: boolean;
}

export interface IWebService {
    GetRoutes(): Array<Route>;
}


export class IWebServiceDeps {
    AppConfig: any;
}

export interface AuthTokenObject {
    username: string,
    userid: number
}